refactored control converter to a base class
