+++ backends/bmv2/backend.cpp
@@ -19,7 +19,6 @@ limitations under the License.
 #include "control.h"
 #include "deparser.h"
 #include "expression.h"
-#include "extern.h"
 #include "globals.h"
 #include "frontends/common/constantFolding.h"
 #include "frontends/p4/evaluator/evaluator.h"
@@ -30,6 +29,7 @@ limitations under the License.
 #include "midend/actionSynthesis.h"
 #include "midend/orderArguments.h"
 #include "midend/removeLeftSlices.h"
+
 #include "lower.h"
 #include "header.h"
 #include "parser.h"
@@ -38,47 +38,6 @@ limitations under the License.
 #include "simpleSwitch.h"
 
 namespace BMV2 {
-
-/**
-This class implements a policy suitable for the SynthesizeActions pass.
-The policy is: do not synthesize actions for the controls whose names
-are in the specified set.
-For example, we expect that the code in the deparser will not use any
-tables or actions.
-*/
-class SkipControls : public P4::ActionSynthesisPolicy {
-    // set of controls where actions are not synthesized
-    const std::set<cstring> *skip;
-
- public:
-    explicit SkipControls(const std::set<cstring> *skip) : skip(skip) { CHECK_NULL(skip); }
-    bool convert(const IR::P4Control* control) const {
-        if (skip->find(control->name) != skip->end())
-            return false;
-        return true;
-    }
-};
-
-/**
-This class implements a policy suitable for the RemoveComplexExpression pass.
-The policy is: only remove complex expression for the controls whose names
-are in the specified set.
-For example, we expect that the code in ingress and egress will have complex
-expression removed.
-*/
-class ProcessControls : public BMV2::RemoveComplexExpressionsPolicy {
-    const std::set<cstring> *process;
-
- public:
-    explicit ProcessControls(const std::set<cstring> *process) : process(process) {
-        CHECK_NULL(process);
-    }
-    bool convert(const IR::P4Control* control) const {
-        if (process->find(control->name) != process->end())
-            return true;
-        return false;
-    }
-};
 
 cstring Backend::jsonAssignment(const IR::Type* type, bool inParser) {
     if (!inParser && type->is<IR::Type_Varbits>())
@@ -100,367 +59,6 @@ cstring Backend::jsonAssignment(const IR::Type* type, bool inParser) {
         return "set";
     else
         return "assign";
-}
-
-/**
-This pass adds @name annotations to all fields of the user metadata
-structure so that they do not clash with fields of the headers
-structure.  This is necessary because both of them become global
-objects in the output json.
-*/
-class RenameUserMetadata : public Transform {
-    P4::ReferenceMap* refMap;
-    const IR::Type_Struct* userMetaType;
-    // Used as a prefix for the fields of the userMetadata structure
-    // and also as a name for the userMetadata type clone.
-    cstring namePrefix;
-
- public:
-    RenameUserMetadata(P4::ReferenceMap* refMap,
-        const IR::Type_Struct* userMetaType, cstring namePrefix):
-            refMap(refMap), userMetaType(userMetaType), namePrefix(namePrefix)
-    { setName("RenameUserMetadata"); CHECK_NULL(refMap); }
-
-    const IR::Node* postorder(IR::Type_Struct* type) override {
-        // Clone the user metadata type.  We want this type to be used
-        // only for parameters.  For any other variables we will used
-        // the clone we create.
-        if (userMetaType != getOriginal())
-            return type;
-
-        auto vec = new IR::IndexedVector<IR::Node>();
-        LOG2("Creating clone" << getOriginal());
-        auto clone = type->clone();
-        clone->name = namePrefix;
-        vec->push_back(clone);
-
-        // Rename all fields
-        IR::IndexedVector<IR::StructField> fields;
-        for (auto f : type->fields) {
-            auto anno = f->getAnnotation(IR::Annotation::nameAnnotation);
-            cstring suffix = "";
-            if (anno != nullptr)
-                suffix = IR::Annotation::getName(anno);
-            if (suffix.startsWith(".")) {
-                // We can't change the name of this field.
-                // Hopefully the user knows what they are doing.
-                fields.push_back(f->clone());
-                continue;
-            }
-
-            if (!suffix.isNullOrEmpty())
-                suffix = cstring(".") + suffix;
-            else
-                suffix = cstring(".") + f->name;
-            cstring newName = namePrefix + suffix;
-            auto stringLit = new IR::StringLiteral(newName);
-            LOG2("Renaming " << f << " to " << newName);
-            auto annos = f->annotations->addOrReplace(
-                IR::Annotation::nameAnnotation, stringLit);
-            auto field = new IR::StructField(f->srcInfo, f->name, annos, f->type);
-            fields.push_back(field);
-        }
-
-        auto annotated = new IR::Type_Struct(
-            type->srcInfo, type->name, type->annotations, std::move(fields));
-        vec->push_back(annotated);
-        return vec;
-    }
-
-    const IR::Node* preorder(IR::Type_Name* type) override {
-        // Find any reference to the user metadata type that is used
-        // (but not for parameters or the package instantiation)
-        // and replace it with the cloned type.
-        if (!findContext<IR::Declaration_Variable>())
-            return type;
-        auto decl = refMap->getDeclaration(type->path);
-        if (decl == userMetaType)
-            type->path = new IR::Path(type->path->srcInfo, IR::ID(type->path->srcInfo, namePrefix));
-        LOG2("Replacing reference with " << type);
-        return type;
-    }
-};
-
-class BuildResourceMap : public Inspector {
- public:
-    Backend* backend;
-
-    BuildResourceMap(Backend *backend) : backend(backend) {};
-
-    bool preorder(const IR::ControlBlock* control) override {
-        backend->resourceMap.emplace(control->container, control);
-        for (auto cv : control->constantValue) {
-            backend->resourceMap.emplace(cv.first, cv.second);
-        }
-
-        for (auto c : control->container->controlLocals) {
-            if (c->is<IR::InstantiatedBlock>()) {
-                backend->resourceMap.emplace(c, control->getValue(c));
-            }
-        }
-        return false;
-    }
-
-    bool preorder(const IR::ParserBlock* parser) override {
-        backend->resourceMap.emplace(parser->container, parser);
-        for (auto cv : parser->constantValue) {
-            backend->resourceMap.emplace(cv.first, cv.second);
-            if (cv.second->is<IR::Block>()) {
-                visit(cv.second->getNode());
-            }
-        }
-
-        for (auto c : parser->container->parserLocals) {
-            if (c->is<IR::InstantiatedBlock>()) {
-                backend->resourceMap.emplace(c, parser->getValue(c));
-            }
-        }
-        return false;
-    }
-
-    bool preorder(const IR::TableBlock* table) override {
-        backend->resourceMap.emplace(table->container, table);
-        for (auto cv : table->constantValue) {
-            backend->resourceMap.emplace(cv.first, cv.second);
-            if (cv.second->is<IR::Block>()) {
-                visit(cv.second->getNode());
-            }
-        }
-        return false;
-    }
-
-    bool preorder(const IR::PackageBlock* package) override {
-        for (auto cv : package->constantValue) {
-            if (cv.second->is<IR::Block>()) {
-                visit(cv.second->getNode());
-            }
-        }
-        return false;
-    }
-
-    bool preorder(const IR::ToplevelBlock* tlb) override {
-        auto package = tlb->getMain();
-        visit(package);
-        return false;
-    }
-};
-
-class ExtractMatchKind : public Inspector {
- public:
-    Backend *backend;
-    ExtractMatchKind(Backend* backend) : backend(backend) {}
-    bool preorder(const IR::Declaration_MatchKind* kind) override {
-        for (auto member : kind->members) {
-            backend->match_kinds.insert(member->name);
-        }
-        return false;
-    }
-};
-
-void
-Backend::convert_simple_switch(const IR::ToplevelBlock* tlb, BMV2Options& options) {
-    CHECK_NULL(tlb);
-    auto evaluator = new P4::EvaluatorPass(refMap, typeMap);
-    if (tlb->getMain() == nullptr)
-        return;  // no main
-
-    /// Declaration which introduces the user metadata.
-    /// We expect this to be a struct type.
-    const IR::Type_Struct* userMetaType = nullptr;
-    cstring userMetaName = refMap->newName("userMetadata");
-
-    simpleSwitch->setPipelineControls(tlb, &pipeline_controls, &pipeline_namemap);
-    simpleSwitch->setNonPipelineControls(tlb, &non_pipeline_controls);
-    simpleSwitch->setComputeChecksumControls(tlb, &compute_checksum_controls);
-    simpleSwitch->setVerifyChecksumControls(tlb, &verify_checksum_controls);
-    simpleSwitch->setDeparserControls(tlb, &deparser_controls);
-    if (::errorCount() > 0)
-        return;
-
-    // Find the user metadata declaration
-    auto parser = simpleSwitch->getParser(tlb);
-    auto params = parser->getApplyParameters();
-    BUG_CHECK(params->size() == 4, "%1%: expected 4 parameters", parser);
-    auto metaParam = params->parameters.at(2);
-    auto paramType = metaParam->type;
-    if (!paramType->is<IR::Type_Name>()) {
-        ::error("%1%: expected the user metadata type to be a struct", paramType);
-        return;
-    }
-    auto decl = refMap->getDeclaration(paramType->to<IR::Type_Name>()->path);
-    if (!decl->is<IR::Type_Struct>()) {
-        ::error("%1%: expected the user metadata type to be a struct", paramType);
-        return;
-    }
-    userMetaType = decl->to<IR::Type_Struct>();
-    LOG2("User metadata type is " << userMetaType);
-
-    // These passes are logically bmv2-specific
-    PassManager simplify = {
-        new ExtractMatchKind(this),
-        new RenameUserMetadata(refMap, userMetaType, userMetaName),
-        new P4::ClearTypeMap(typeMap),  // because the user metadata type has changed
-        new P4::OrderArguments(refMap, typeMap),
-        new P4::SynthesizeActions(refMap, typeMap, new SkipControls(&non_pipeline_controls)),
-        new P4::MoveActionsToTables(refMap, typeMap),
-        new P4::TypeChecking(refMap, typeMap),
-        new P4::SimplifyControlFlow(refMap, typeMap),
-        new LowerExpressions(typeMap),
-        new P4::ConstantFolding(refMap, typeMap, false),
-        new P4::TypeChecking(refMap, typeMap),
-        new RemoveComplexExpressions(refMap, typeMap, new ProcessControls(&pipeline_controls)),
-        new P4::SimplifyControlFlow(refMap, typeMap),
-        new P4::RemoveAllUnusedDeclarations(refMap),
-        new DiscoverStructure(&structure),
-        evaluator,
-    };
-    tlb->getProgram()->apply(simplify);
-
-    toplevel = evaluator->getToplevelBlock();
-    toplevel->apply(*new BuildResourceMap(this));
-
-    jsonTop.emplace("program", options.file);
-    jsonTop.emplace("__meta__", json->meta);
-    jsonTop.emplace("header_types", json->header_types);
-    jsonTop.emplace("headers", json->headers);
-    jsonTop.emplace("header_stacks", json->header_stacks);
-    jsonTop.emplace("header_union_types", json->header_union_types);
-    jsonTop.emplace("header_unions", json->header_unions);
-    jsonTop.emplace("header_union_stacks", json->header_union_stacks);
-    field_lists = mkArrayField(&jsonTop, "field_lists");
-    // field list and learn list ids in bmv2 are not consistent with ids for
-    // other objects: they need to start at 1 (not 0) since the id is also used
-    // as a "flag" to indicate that a certain simple_switch primitive has been
-    // called (e.g. resubmit or generate_digest)
-    BMV2::nextId("field_lists");
-    jsonTop.emplace("errors", json->errors);
-    jsonTop.emplace("enums", json->enums);
-    jsonTop.emplace("parsers", json->parsers);
-    jsonTop.emplace("parse_vsets", json->parse_vsets);
-    jsonTop.emplace("deparsers", json->deparsers);
-    meter_arrays = mkArrayField(&jsonTop, "meter_arrays");
-    counters = mkArrayField(&jsonTop, "counter_arrays");
-    register_arrays = mkArrayField(&jsonTop, "register_arrays");
-    jsonTop.emplace("calculations", json->calculations);
-    learn_lists = mkArrayField(&jsonTop, "learn_lists");
-    BMV2::nextId("learn_lists");
-    jsonTop.emplace("actions", json->actions);
-    jsonTop.emplace("pipelines", json->pipelines);
-    jsonTop.emplace("checksums", json->checksums);
-    force_arith = mkArrayField(&jsonTop, "force_arith");
-    jsonTop.emplace("extern_instances", json->externs);
-    jsonTop.emplace("field_aliases", json->field_aliases);
-
-    json->add_program_info(options.file);
-    json->add_meta_info();
-
-    // convert all enums to json
-    for (const auto &pEnum : *enumMap) {
-        auto name = pEnum.first->getName();
-        for (const auto &pEntry : *pEnum.second) {
-            json->add_enum(name, pEntry.first, pEntry.second);
-        }
-    }
-    if (::errorCount() > 0)
-        return;
-
-    /// generate error types
-    for (const auto &p : structure.errorCodesMap) {
-        auto name = p.first->toString();
-        auto type = p.second;
-        json->add_error(name, type);
-    }
-
-    cstring scalarsName = refMap->newName("scalars");
-
-    // This visitor is used in multiple passes to convert expression to json
-    conv = new P4V1::SimpleSwitchExpressionConverter(refMap, typeMap, &structure,
-                                                     scalarsName, toplevel);
-
-    PassManager codegen_passes = {
-        new HeaderConverter(refMap, typeMap, &structure, json, scalarsName),
-        new ParserConverter(refMap, typeMap, json, conv),
-        new VisitFunctor([this]() { simpleSwitch->createActions(); }),
-        new ControlConverter(this, options.emitExterns),
-        new ChecksumConverter(this),
-        new DeparserConverter(this),
-    };
-
-    codegen_passes.setName("CodeGen");
-    CHECK_NULL(toplevel);
-    auto program = toplevel->getProgram();
-    program->apply(codegen_passes);
-
-    (void)toplevel->apply(ConvertGlobals(this));
-}
-
-void Backend::convert_portable_switch(const IR::ToplevelBlock* tlb, BMV2Options& options) {
-    CHECK_NULL(tlb);
-    P4::PsaProgramStructure structure(refMap, typeMap);
-
-    auto parsePsaArch = new P4::ParsePsaArchitecture(&structure);
-    auto main = tlb->getMain();
-    if (!main) return;
-    main->apply(*parsePsaArch);
-
-    auto evaluator = new P4::EvaluatorPass(refMap, typeMap);
-    auto program = tlb->getProgram();
-    PassManager toJson = {
-        // new RenameUserMetadata(refMap, userMetaType, userMetaName),
-        new P4::ClearTypeMap(typeMap),  // because the user metadata type has changed
-        new P4::SynthesizeActions(refMap, typeMap, new SkipControls(&non_pipeline_controls)),
-        new P4::MoveActionsToTables(refMap, typeMap),
-        new P4::TypeChecking(refMap, typeMap),
-        new P4::SimplifyControlFlow(refMap, typeMap),
-        new LowerExpressions(typeMap),
-        new P4::ConstantFolding(refMap, typeMap, false),
-        new P4::TypeChecking(refMap, typeMap),
-        new RemoveComplexExpressions(refMap, typeMap, new ProcessControls(&pipeline_controls)),
-        new P4::SimplifyControlFlow(refMap, typeMap),
-        new P4::RemoveAllUnusedDeclarations(refMap),
-        evaluator,
-        new VisitFunctor([this, evaluator]() { toplevel = evaluator->getToplevelBlock(); }),
-        new BMV2::DiscoverStructure(&structure),
-        new P4::InspectPsaProgram(refMap, typeMap, &structure),
-        new P4::ConvertToJson(&structure),
-    };
-    program->apply(toJson);
-
-    auto json = structure.getJson();
-    jsonTop.emplace("program", options.file);
-    jsonTop.emplace("__meta__", json->meta);
-    jsonTop.emplace("header_types", json->header_types);
-    jsonTop.emplace("headers", json->headers);
-    jsonTop.emplace("header_stacks", json->header_stacks);
-    jsonTop.emplace("header_union_types", json->header_union_types);
-    jsonTop.emplace("header_unions", json->header_unions);
-    jsonTop.emplace("header_union_stacks", json->header_union_stacks);
-    field_lists = mkArrayField(&jsonTop, "field_lists");
-    // field list and learn list ids in bmv2 are not consistent with ids for
-    // other objects: they need to start at 1 (not 0) since the id is also used
-    // as a "flag" to indicate that a certain simple_switch primitive has been
-    // called (e.g. resubmit or generate_digest)
-    BMV2::nextId("field_lists");
-    jsonTop.emplace("errors", json->errors);
-    jsonTop.emplace("enums", json->enums);
-    jsonTop.emplace("parsers", json->parsers);
-    jsonTop.emplace("parse_vsets", json->parse_vsets);
-    jsonTop.emplace("deparsers", json->deparsers);
-    meter_arrays = mkArrayField(&jsonTop, "meter_arrays");
-    counters = mkArrayField(&jsonTop, "counter_arrays");
-    register_arrays = mkArrayField(&jsonTop, "register_arrays");
-    jsonTop.emplace("calculations", json->calculations);
-    learn_lists = mkArrayField(&jsonTop, "learn_lists");
-    BMV2::nextId("learn_lists");
-    jsonTop.emplace("actions", json->actions);
-    jsonTop.emplace("pipelines", json->pipelines);
-    jsonTop.emplace("checksums", json->checksums);
-    force_arith = mkArrayField(&jsonTop, "force_arith");
-    jsonTop.emplace("extern_instances", json->externs);
-    jsonTop.emplace("field_aliases", json->field_aliases);
-
-    json->add_program_info(options.file);
-    json->add_meta_info();
 }
 
 }  // namespace BMV2
